@@ -25,7 +25,7 @@
     }
   },
   "require": {
-    "php": ">=7.3",
+    "php": "^8.0",
     "nesbot/carbon": "~2",
     "illuminate/database": "^8.0|^9.0|^10.0",
     "illuminate/support": "^8.0|^9.0|^10.0",
