@@ -1,4 +1,6 @@
-<?php namespace Models;
+<?php
+
+namespace Models;
 
 use Jwohlfert23\LaravelApiQuery\BuildQueryFromRequest;
 use Jwohlfert23\LaravelApiQuery\Searchable;
@@ -10,14 +12,14 @@ class Model extends \Illuminate\Database\Eloquent\Model
     protected $casts = [
         'bool' => 'bool',
         'date' => 'date',
-        'custom_date' => 'date:Y-m'
+        'custom_date' => 'date:Y-m',
     ];
 
     protected $searchable = [
         'columns' => [
             'name' => 10,
-            'id' => 10
-        ]
+            'id' => 10,
+        ],
     ];
 
     public function related()
